# Initialize sub-query only once