@ironage 2023-02-22
I added a benchmark, feel free to modify it or merge as is.

SubqueryStrings (MemOnly, EncryptionOff) Release

  | Req runs: | Min | Max | Median | Avg | Stddev
-- | -- | -- | -- | -- | -- | --
Master | 5 | 122ms | 125ms | 122ms | 123ms | 1ms
This branch | 22 | 22ms | 22ms | 22ms | 22ms | 164us